# docs/openapi/proximadb-openapi.yaml
components:
...
  schemas:
...
    ErrorBody:
      description: Inner body of [`ErrorResponse`].
      properties:
        code:
          description: HTTP status code.
          format: int32
          type: integer
        details:
          description: Optional structured context (e.g. migration hints).
          type:
          - object
          - 'null'
        message:
          type: string
        request_id:
          description: Correlation id (matches the X-Request-ID header).
          type:
          - string
          - 'null'
        type:
          description: Stable machine-readable error code (snake_case).
          example: collection_not_found
          type: string
      required:
      - type
      - message
      - code
      type: object
    ErrorResponse:
      description: |-
        Canonical ProximaDB error envelope (`{ error: { type, message, code } }`).

        `request_id` (also returned in the `X-Request-ID` response header) is present
        whenever the request passed through the request-id middleware; quote it in
        bug reports.
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      required:
      - error
      type: object
    ExplainQueryRequest:
      properties:
        collection:
          type:
          - string
          - 'null'
        language:
          $ref: '#/components/schemas/QueryLanguage'
        query:
          minLength: 1
          type: string
      required:
      - language
      - query
      type: object
...
    QueryLanguage:
      enum:
      - uql
      - aql
      - federated
      type: string
    QueryRequest:
      properties:
        collection:
          type:
          - string
          - 'null'
        language:
          $ref: '#/components/schemas/QueryLanguage'
        limit:
          format: int32
          minimum: 0
          type:
          - integer
          - 'null'
        parameters:
          items:
            type: object
          type:
          - array
          - 'null'
        query:
          minLength: 1
          type: string
      required:
      - language
      - query
      type: object
    QueryResponse:
      description: |-
        Query facade result. Implementations return records, total_count, metrics,
        plan, or diagnostics depending on language and endpoint, so the body is a
        free-form JSON object.
      type: object
...
    SqlRequest:
      description: |-
        One SQL statement executed through the shared SQL authority.

        Parameter binding is intentionally not advertised yet: the relational
        execution port currently has no typed binding contract. Callers must not be
        offered a field that an adapter would silently ignore or interpolate.
      properties:
        collection:
          description: Optional collection context used by vector/graph SQL extensions.
          type:
          - string
          - 'null'
        query:
          maxLength: 1048576
          minLength: 1
          type: string
        timeout_ms:
          description: Per-request deadline. Defaults to 30 seconds and is capped at 5 minutes.
          format: int64
          maximum: 300000
          minimum: 1
          type:
          - integer
          - 'null'
      required:
      - query
      type: object
    SqlResponse:
      description: |-
        Stable JSON envelope for SQL reads and writes.

        For reads, `rows_returned` is the row count. For DDL/DML, the shared SQL
        authority reports the affected count in `rows_returned` and `rows` is empty.
      properties:
        column_types:
          items:
            type: string
          type: array
        columns:
          items:
            type: string
          type: array
        execution_time_ms:
          format: int64
          minimum: 0
          type: integer
        request_id:
          type: string
        rows:
          items:
            type: object
          type: array
        rows_returned:
          format: int64
          minimum: 0
          type: integer
        rows_scanned:
          format: int64
          minimum: 0
          type: integer
      required:
      - rows
      - columns
      - column_types
      - rows_returned
      - rows_scanned
      - execution_time_ms
      - request_id
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    email: vijay@anvaiops.com
    name: Vijaykumar Singh
  description: SDK-facing ProximaDB REST contract. This specification is v2-first and centers on ProximaRecord, typed schemas, record search, and collection lifecycle operations. Legacy v1 compatibility routes are intentionally not part of this publishable SDK surface.
  license:
    identifier: Apache-2.0
    name: Apache-2.0
  title: ProximaDB REST API
  version: 0.2.0
openapi: 3.1.0
paths:
...
  /api/v2/query:
    post:
      description: |-
        The canonical unified query surface: UQL (unified multi-modal), federated SQL
        extensions, and AQL. This is NOT plain SQL-over-REST and is NOT deprecated:
        UQL/AQL are non-SQL languages that pgwire and `/api/v2/sql` cannot serve, so
        this endpoint remains their canonical home. Authenticated programmatic SQL
        is available through both gRPC `ExecuteQuery` and `/api/v2/sql`; neither is a
        replacement for this UQL/federated surface.
      operationId: executeQuery
      parameters:
      - description: Optional explicit tenant selector. Applied only when there is no authenticated tenant context — a JWT tenant claim takes precedence, and a header that disagrees with the authenticated tenant is rejected. Absent ⇒ the default tenant. Tenant isolation is structural on the server; this header only selects the tenant.
        in: header
        name: X-Tenant-ID
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
          description: Query result.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
      summary: Execute AQL or UQL through the shared query facade.
      tags:
      - Query
  /api/v2/query/explain:
    post:
      operationId: explainQuery
      parameters:
      - description: Optional explicit tenant selector. Applied only when there is no authenticated tenant context — a JWT tenant claim takes precedence, and a header that disagrees with the authenticated tenant is rejected. Absent ⇒ the default tenant. Tenant isolation is structural on the server; this header only selects the tenant.
        in: header
        name: X-Tenant-ID
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExplainQueryRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
          description: Query plan and lowering details.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid request.
      summary: Explain an AQL or UQL query through the shared query facade.
      tags:
      - Query
  /api/v2/sql:
    post:
      description: |-
        The required foundation identity is inserted by the root tenant middleware
        after authentication. A router mounted without that middleware fails closed
        at extraction instead of constructing an anonymous authorization carrier.
      operationId: executeSql
      parameters:
      - description: Optional explicit tenant selector. Applied only when there is no authenticated tenant context — a JWT tenant claim takes precedence, and a header that disagrees with the authenticated tenant is rejected. Absent ⇒ the default tenant. Tenant isolation is structural on the server; this header only selects the tenant.
        in: header
        name: X-Tenant-ID
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SqlRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SqlResponse'
          description: SQL result or write count.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid SQL request.
        '408':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Query deadline exceeded.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Write lock conflict.
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: SQL execution failure.
      summary: Execute one authenticated SQL statement.
      tags:
      - SQL
...
security:
- bearerAuth: []
servers:
- description: Local ProximaDB server
  url: http://localhost:5678
...